# public/openapi.yaml
openapi: 3.0.3
info:
  title: Seva Sansaar API
  version: "1.0.0"
  description: |
    HTTP API for Seva Sansaar (bookings, businesses, reviews, owner portal, admin).
    Admin routes require header `x-api-key` matching `ADMIN_API_KEY` (and `NEXT_PUBLIC_ADMIN_API_KEY` on the client).
servers:
  - url: /
    description: Same origin as the Next.js app
...
paths:
...
  /api/reviews:
    post:
      tags: [Reviews]
      summary: Submit review (auth)
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT